fix windows CI

+++ test/e2e/e2e_utils_test.cpp
@@ -35,7 +35,7 @@ TEST_F(E2E_Utils, submit_vbk) {
                 this->mempool.getInFlightMap<VbkBlock>().size(),
             1);
 
-  EXPECT_EQ(this->e2e.getStats().created_vbk, 1);
+  EXPECT_EQ(this->e2e.getStats().created_vbk, (uint32_t)1);
 }
 
 TEST_F(E2E_Utils, submit_vtb) {
@@ -57,10 +57,10 @@ TEST_F(E2E_Utils, submit_vtb) {
                 this->mempool.getInFlightMap<VTB>().size(),
             1);
 
-  EXPECT_EQ(this->e2e.getStats().created_vbk, 1);
-  EXPECT_EQ(this->e2e.getStats().created_vbk_pop_tx, 1);
-  EXPECT_EQ(this->e2e.getStats().created_btc, 1);
-  EXPECT_EQ(this->e2e.getStats().created_btc_tx, 1);
+  EXPECT_EQ(this->e2e.getStats().created_vbk, (uint32_t)1);
+  EXPECT_EQ(this->e2e.getStats().created_vbk_pop_tx, (uint32_t)1);
+  EXPECT_EQ(this->e2e.getStats().created_btc, (uint32_t)1);
+  EXPECT_EQ(this->e2e.getStats().created_btc_tx, (uint32_t)1);
 }
 
 TEST_F(E2E_Utils, submit_atv) {
@@ -80,8 +80,8 @@ TEST_F(E2E_Utils, submit_atv) {
                 this->mempool.getInFlightMap<ATV>().size(),
             1);
 
-  EXPECT_EQ(this->e2e.getStats().created_vbk, 1);
-  EXPECT_EQ(this->e2e.getStats().created_vbk_tx, 1);
+  EXPECT_EQ(this->e2e.getStats().created_vbk, (uint32_t)1);
+  EXPECT_EQ(this->e2e.getStats().created_vbk_tx, (uint32_t)1);
 }
 
 TEST_F(E2E_Utils, submit_alt) {
@@ -89,12 +89,12 @@ TEST_F(E2E_Utils, submit_alt) {
 
   this->e2e.createAction(CreateOption::CREATE_ALT, fork, this->alttree);
 
-  EXPECT_EQ(this->alttree.getAllBlocks().size(), 1);
+  EXPECT_EQ(this->alttree.getAllBlocks().size(), (uint32_t)1);
 
   this->e2e.submitAction(
       SubmitOption::SUBMIT_ALT, this->mempool, this->alttree);
 
-  EXPECT_EQ(this->alttree.getAllBlocks().size(), 2);
+  EXPECT_EQ(this->alttree.getAllBlocks().size(), (size_t)2);
 
-  EXPECT_EQ(this->e2e.getStats().created_alt, 1);
+  EXPECT_EQ(this->e2e.getStats().created_alt, (uint32_t)1);
 }
